requirements.txt
# Web Framework and API Dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
starlette==0.27.0
python-jose[cryptography]==3.3.0  # For JWT handling
bcrypt==4.0.1  # For password hashing

# Database and ORM
sqlalchemy==2.0.23
alembic==1.13.0
asyncpg==0.29.0  # PostgreSQL async driver
psycopg2-binary==2.9.9  # PostgreSQL sync driver
supabase==2.0.0  # Supabase client
python-dotenv==1.0.0  # Environment variable management

# AI and ML Dependencies
sentence-transformers==2.2.2  # For semantic text matching
scikit-learn==1.3.2  # For ML utilities
torch==2.1.0  # PyTorch for ML models
transformers==4.34.0  # Hugging Face transformers
numpy==1.24.3  # Numerical computing
pandas==2.1.1  # Data manipulation

# Audio Processing
edge-tts==6.1.9  # Microsoft Edge TTS
pydub==0.25.1  # Audio file manipulation
aiofiles==23.2.1  # Async file operations
soundfile==0.12.1  # Audio file reading/writing

# File Processing
PyPDF2==3.0.1  # PDF file handling
python-magic==0.4.27  # File type detection
pillow==10.0.1  # Image processing
python-docx==1.0.0  # Word document processing

# Testing and Development
pytest==7.4.2
pytest-asyncio==0.21.1
httpx==0.25.0  # Async HTTP client
pytest-cov==4.1.0  # Coverage reporting
black==23.9.1  # Code formatting
isort==5.12.0  # Import sorting
flake8==6.1.0  # Code linting

# Monitoring and Logging
prometheus-client==0.17.1  # Metrics
python-json-logger==2.0.7  # JSON logging
structlog==23.1.0  # Structured logging

# Additional Utilities
python-dateutil==2.8.2  # Date utilities
pytz==2023.3.post1  # Timezone support
tqdm==4.66.1  # Progress bars
tenacity==8.2.3  # Retry logic
ujson==5.8.0  # Fast JSON
orjson==3.9.7  # Even faster JSON
email-validator==2.0.0  # Email validation
jinja2==3.1.2  # Template engine


openai>=1.0.0
python-dotenv>=0.19.0

sentence-transformers>=2.2.0
scikit-learn>=1.0.0

numpy>=1.21.0
torch>=1.9.0

requests>=2.25.0
